# docs/swagger.yaml
definitions:
...
  handler.MemberPage:
    properties:
      items:
        items:
          $ref: '#/definitions/model.ProjectUser'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        format: int64
        type: integer
    type: object
  handler.ProjectDocumentPage:
    properties:
      items:
        items:
          $ref: '#/definitions/model.ProjectDocument'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        format: int64
        type: integer
    type: object
...
  model.ColumnType:
    enum:
    - normal
    - done
    - blocked
    type: string
    x-enum-varnames:
    - ColumnNormal
    - ColumnDone
    - ColumnBlocked
...
  model.KanbanColumn:
    properties:
      boardId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      key:
        type: string
      name:
        type: string
      position:
        type: integer
      type:
        $ref: '#/definitions/model.ColumnType'
    type: object
  model.KanbanColumnRequest:
    properties:
      key:
        type: string
      name:
        type: string
      position:
        type: integer
      type:
        $ref: '#/definitions/model.ColumnType'
    type: object
...
  model.ProjectDocument:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    type: object
  model.ProjectDocumentRequest:
    properties:
      content:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  model.ProjectRole:
    enum:
    - owner
    - admin
    - developer
    - guest
    type: string
    x-enum-varnames:
    - Owner
    - Admin
    - Developer
    - Guest
  model.ProjectUser:
    properties:
      isStarred:
        type: boolean
      joinedAt:
        type: string
      projectId:
        type: string
      role:
        $ref: '#/definitions/model.ProjectRole'
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  model.UpdateProjectUserRequest:
    properties:
      role:
        $ref: '#/definitions/model.ProjectRole'
    type: object
  model.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/inri/kanban/{kanbanId}/columns:
    get:
      consumes:
      - application/json
      description: Returns all columns for a given kanban board
      parameters:
      - description: Kanban board ID
        in: path
        name: kanbanId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.KanbanColumn'
            type: array
      security:
      - GatewayAuth: []
      tags:
      - Kanban Columns
    post:
      consumes:
      - application/json
      description: Creates a new column in the kanban board
      parameters:
      - description: Kanban board ID
        in: path
        name: kanbanId
        required: true
        type: string
      - description: Column data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KanbanColumnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.KanbanColumn'
      security:
      - GatewayAuth: []
      tags:
      - Kanban Columns
  /v1/inri/kanban/{kanbanId}/columns/{columnId}:
    delete:
      consumes:
      - application/json
      description: Deletes a column from the kanban board
      parameters:
      - description: Kanban board ID
        in: path
        name: kanbanId
        required: true
        type: string
      - description: Column ID
        in: path
        name: columnId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - Kanban Columns
    get:
      consumes:
      - application/json
      description: Returns a specific column by ID
      parameters:
      - description: Kanban board ID
        in: path
        name: kanbanId
        required: true
        type: string
      - description: Column ID
        in: path
        name: columnId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KanbanColumn'
      security:
      - GatewayAuth: []
      tags:
      - Kanban Columns
    put:
      consumes:
      - application/json
      description: Updates an existing column
      parameters:
      - description: Kanban board ID
        in: path
        name: kanbanId
        required: true
        type: string
      - description: Column ID
        in: path
        name: columnId
        required: true
        type: string
      - description: Updated column data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.KanbanColumnRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.KanbanColumn'
      security:
      - GatewayAuth: []
      tags:
      - Kanban Columns
...
  /v1/inri/projects/{projectId}/documents:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of project documents for a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: pageSize
        type: integer
      - description: RSQL filter query
        in: query
        name: filter
        type: string
      - description: Sort query
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ProjectDocumentPage'
      security:
      - GatewayAuth: []
      tags:
      - ProjectDocuments
    post:
      consumes:
      - application/json
      description: Creates a new project document under the project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Document data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/model.ProjectDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ProjectDocument'
      security:
      - GatewayAuth: []
      tags:
      - ProjectDocuments
  /v1/inri/projects/{projectId}/documents/{documentId}:
    delete:
      consumes:
      - application/json
      description: Deletes the project document
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - ProjectDocuments
    get:
      consumes:
      - application/json
      description: Returns a project document by its ID within the project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectDocument'
      security:
      - GatewayAuth: []
      tags:
      - ProjectDocuments
    put:
      consumes:
      - application/json
      description: Updates an existing project document
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Document ID
        in: path
        name: documentId
        required: true
        type: string
      - description: Document data
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/model.ProjectDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectDocument'
      security:
      - GatewayAuth: []
      tags:
      - ProjectDocuments
...
  /v1/inri/projects/{projectId}/members:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of all project members
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: pageSize
        type: integer
      - description: RSQL filter query
        in: query
        name: filter
        type: string
      - description: Sort query
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MemberPage'
      security:
      - GatewayAuth: []
      tags:
      - Projects
  /v1/inri/projects/{projectId}/members/{memberId}:
    delete:
      consumes:
      - application/json
      description: Removes a member from a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Member User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - GatewayAuth: []
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates the role of a project member
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Member User ID
        in: path
        name: memberId
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectUser'
      security:
      - GatewayAuth: []
      tags:
      - Projects
...
securityDefinitions:
  GatewayAuth:
    description: Injected by API Gateway. Do not use in production.
    in: header
    name: X-User-ID
    type: apiKey
swagger: "2.0"